ml-api/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2

# ── Emotion model (ViT) ───────────────────────────────────────────────────────
transformers>=4.40.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.24.0

# ── Emission model (ResNet-50) ────────────────────────────────────────────────
# torch / torchvision already listed above

# ── HTTP client (async image fetching for predict-url) ────────────────────────
httpx>=0.27.0
